Rework ApplePay session
