+++ Sources/ProcessOut/Sources/Sessions/ApplePay/ApplePayAuthorizationSession.swift
@@ -9,10 +9,9 @@ import PassKit
 
 protocol ApplePayAuthorizationSession: Sendable {
 
-    /// Begins an Apple Pay authorization.
-    func authorize<T>(
-        request: PKPaymentRequest,
-        didAuthorizePayment: @escaping (PKPayment) async throws -> T,
-        delegate: POApplePayAuthorizationSessionDelegate?
-    ) async throws -> T
+    /// Begins an Apple Pay payment authorization.
+    /// - NOTE: delegate is retained for the duration of presentation.
+    func authorize(
+        request: PKPaymentRequest, delegate: ApplePayAuthorizationSessionDelegate?
+    ) async throws -> PKPayment
 }
