+++ Sources/ProcessOut/Sources/Sessions/ApplePay/DefaultApplePayAuthorizationSession.swift
@@ -16,15 +16,12 @@ final class DefaultApplePayAuthorizationSession: ApplePayAuthorizationSession {
 
     // MARK: - ApplePayAuthorizationSession
 
-    func authorize<T>(
+    func authorize(
         request: PKPaymentRequest,
-        didAuthorizePayment: @escaping (PKPayment) async throws -> T,
-        delegate: POApplePayAuthorizationSessionDelegate? = nil
-    ) async throws -> T {
+        delegate: ApplePayAuthorizationSessionDelegate?
+    ) async throws -> PKPayment {
         let controller = PKPaymentAuthorizationController(paymentRequest: request)
-        let coordinator = AuthorizationSessionCoordinator(
-            didAuthorizePayment: didAuthorizePayment, delegate: delegate
-        )
+        let coordinator = ApplePayAuthorizationSessionCoordinator(delegate: delegate)
         controller.delegate = coordinator
         guard await controller.present() else {
             throw POFailure(message: "Unable to present authorization controller.", code: .generic(.mobile))
@@ -37,114 +34,9 @@ final class DefaultApplePayAuthorizationSession: ApplePayAuthorizationSession {
             controller.dismiss()
         }
         await controller.dismiss()
-        guard let processedPayment = coordinator.processedPayment else {
+        guard let payment = coordinator.payment else {
             throw POFailure(message: "Authorization was cancelled.", code: .cancelled)
         }
-        return processedPayment
+        return payment
     }
-}
-
-@MainActor
-private final class AuthorizationSessionCoordinator<ProcessedPayment>
-    : NSObject, PKPaymentAuthorizationControllerDelegate {
-
-    init(
-        didAuthorizePayment: @escaping (PKPayment) async throws -> ProcessedPayment,
-        delegate: POApplePayAuthorizationSessionDelegate?
-    ) {
-        self.delegate = delegate
-        self.didAuthorizePayment = didAuthorizePayment
-        didFinish = false
-        super.init()
-    }
-
-    /// Processed payment information if any.
-    var processedPayment: ProcessedPayment?
-
-    func setContinuation(continuation: CheckedContinuation<Void, Never>) {
-        if didFinish {
-            continuation.resume()
-        } else {
-            self.continuation = continuation
-        }
-    }
-
-    // MARK: - PKPaymentAuthorizationControllerDelegate
-
-    func paymentAuthorizationControllerDidFinish(_ controller: PKPaymentAuthorizationController) {
-        didFinish = true
-        continuation?.resume()
-        continuation = nil
-    }
-
-    func paymentAuthorizationController(
-        _ controller: PKPaymentAuthorizationController, didAuthorizePayment payment: PKPayment
-    ) async -> PKPaymentAuthorizationResult {
-        do {
-            let processedPayment = try await didAuthorizePayment(payment)
-            // swiftlint:disable:next line_length
-            let result = await delegate?.applePayAuthorizationSession(didAuthorizePayment: payment) ?? .init(status: .success, errors: nil)
-            if case .success = result.status {
-                self.processedPayment = processedPayment
-            }
-            return result
-        } catch {
-            // todo(andrii-vysotskyi): map error
-            return .init(status: .failure, errors: nil)
-        }
-    }
-
-    @MainActor
-    func paymentAuthorizationControllerWillAuthorizePayment(_ controller: PKPaymentAuthorizationController) {
-        // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    @available(iOS 14.0, *)
-    func paymentAuthorizationControllerDidRequestMerchantSessionUpdate(
-        controller: PKPaymentAuthorizationController
-    ) async -> PKPaymentRequestMerchantSessionUpdate {
-        .init() // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    @available(iOS 15.0, *)
-    func paymentAuthorizationController(
-        _ controller: PKPaymentAuthorizationController,
-        didChangeCouponCode couponCode: String
-    ) async -> PKPaymentRequestCouponCodeUpdate {
-        .init() // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    func paymentAuthorizationController(
-        _ controller: PKPaymentAuthorizationController,
-        didSelectShippingMethod shippingMethod: PKShippingMethod
-    ) async -> PKPaymentRequestShippingMethodUpdate {
-        .init() // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    func paymentAuthorizationController(
-        _ controller: PKPaymentAuthorizationController,
-        didSelectShippingContact contact: PKContact
-    ) async -> PKPaymentRequestShippingContactUpdate {
-        .init() // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    func paymentAuthorizationController(
-        _ controller: PKPaymentAuthorizationController,
-        didSelectPaymentMethod paymentMethod: PKPaymentMethod
-    ) async -> PKPaymentRequestPaymentMethodUpdate {
-        .init() // todo(andrii-vysotskyi): forward to delegate
-    }
-
-    @available(iOS 14.0, *)
-    nonisolated func presentationWindow(for controller: PKPaymentAuthorizationController) -> UIWindow? {
-        nil
-    }
-
-    // MARK: - Private Properties
-
-    private let delegate: POApplePayAuthorizationSessionDelegate?
-    private let didAuthorizePayment: (PKPayment) async throws -> ProcessedPayment
-
-    private var continuation: CheckedContinuation<Void, Never>?
-    private var didFinish: Bool
 }
